Adds versioning to Docker image builds
Adds a build argument to the Dockerfile to include the git commit SHA as the version.
This allows for better tracking and identification of specific builds.

The Dockerfile is updated to use build arguments for OS and architecture,
and to inject the version during the build process.

The github workflow is updated to pass the github SHA as the VERSION build argument.